extract overlapping_model fully, find missing headers in some files this way & rework rgba ==
